+++ srcs/executor/handle_cmd_1.c
@@ -6,7 +6,7 @@
 /*   By: achappui <achappui@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 16:57:47 by achappui          #+#    #+#             */
-/*   Updated: 2024/04/17 22:54:58 by achappui         ###   ########.fr       */
+/*   Updated: 2024/04/18 12:56:54 by achappui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
