+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: achappui <achappui@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/23 17:38:15 by achappui          #+#    #+#             */
-/*   Updated: 2024/04/17 22:48:43 by achappui         ###   ########.fr       */
+/*   Updated: 2024/04/18 11:46:10 by achappui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
