+++ srcs/parser/make_tree_3.c
@@ -87,3 +87,25 @@ t_node	*handle_cmd(t_token_list *start, t_token_list *end, t_minishell *info)
 	}
 	return (new_node);
 }
+
+char	*ft_itoa_heredoc(uintptr_t n)
+{
+	uintptr_t	tab;
+	short		i;
+
+	tab = n;
+	i = 2;
+	while ((tab >= 10) && i++)
+		tab /= 10;
+	tab = (long long)malloc((i + 2) * sizeof(char));
+	if (!tab)
+		return (NULL);
+	((char *)tab)[i] = '\0';
+	((char *)tab)[0] = '.';
+	while (n > 0 || (i == 1))
+	{
+		((char *)tab)[--i] = n % 10 + 48;
+		n /= 10;
+	}
+	return ((char *)tab);
+}
